// packages/prettier-plugin-apex/vendor/apex-ast-serializer/typings/jorje.d.ts
export interface CompilationException extends RuntimeException, Locatable {
    "@class": "apex.jorje.services.exception.InternalException" | "apex.jorje.services.exception.ParseException";
    error: string;
}
// ...
export interface Expr {
    "@class": "apex.jorje.data.ast.Expr$ArrayExpr" | "apex.jorje.data.ast.Expr$AssignmentExpr" | "apex.jorje.data.ast.Expr$BinaryExpr" | "apex.jorje.data.ast.Expr$BooleanExpr" | "apex.jorje.data.ast.Expr$CastExpr" | "apex.jorje.data.ast.Expr$ClassRefExpr" | "apex.jorje.data.ast.Expr$InstanceOf" | "apex.jorje.data.ast.Expr$JavaMethodCallExpr" | "apex.jorje.data.ast.Expr$JavaVariableExpr" | "apex.jorje.data.ast.Expr$LiteralExpr" | "apex.jorje.data.ast.Expr$MethodCallExpr" | "apex.jorje.data.ast.Expr$NestedExpr" | "apex.jorje.data.ast.Expr$NewExpr" | "apex.jorje.data.ast.Expr$PackageVersionExpr" | "apex.jorje.data.ast.Expr$PostfixExpr" | "apex.jorje.data.ast.Expr$PrefixExpr" | "apex.jorje.data.ast.Expr$SoqlExpr" | "apex.jorje.data.ast.Expr$SoslExpr" | "apex.jorje.data.ast.Expr$SuperMethodCallExpr" | "apex.jorje.data.ast.Expr$SuperVariableExpr" | "apex.jorje.data.ast.Expr$TernaryExpr" | "apex.jorje.data.ast.Expr$ThisMethodCallExpr" | "apex.jorje.data.ast.Expr$ThisVariableExpr" | "apex.jorje.data.ast.Expr$TriggerVariableExpr" | "apex.jorje.data.ast.Expr$VariableExpr";
}
// ...
export interface IfBlock {
    "@class": "apex.jorje.data.ast.IfBlock";
    expr: Expr;
    loc: Location;
    stmnt: Stmnt;
    "@id"?: string;
    "@reference"?: string;
}
// ...
export interface IndexLocation extends Location {
    "@class": "apex.jorje.data.IndexLocation";
    "@id"?: string;
    "@reference"?: string;
}
// ...
export interface InternalException extends CompilationException {
    "@class": "apex.jorje.services.exception.InternalException";
    "@id"?: string;
    "@reference"?: string;
}
// ...
export interface Locatable {
    "@class": "apex.jorje.data.Identifiers$LocationIdentifier" | "apex.jorje.data.Identifiers$SyntheticIdentifier" | "apex.jorje.data.IndexLocation" | "apex.jorje.data.PositionLocation" | "apex.jorje.services.exception.InternalException" | "apex.jorje.services.exception.ParseException";
    loc: Location;
}

export interface Locatables {
    "@class": "apex.jorje.data.Locatables";
    "@id"?: string;
    "@reference"?: string;
}

export interface Location extends Locatable {
    "@class": "apex.jorje.data.IndexLocation" | "apex.jorje.data.PositionLocation";
    column: number;
    endIndex: number;
    line: number;
    startIndex: number;
}
// ...
export interface ParseException extends CompilationException {
    "@class": "apex.jorje.services.exception.ParseException";
    userError: UserError;
    "@id"?: string;
    "@reference"?: string;
}

export interface ParserOutput {
    "@class": "apex.jorje.semantic.compiler.parser.ParserOutput";
    hiddenTokenMap: [{"@class": string}, HiddenToken][];
    internalErrors: InternalException[];
    parseErrors: ParseException[];
    unit: CompilationUnit;
    "@id"?: string;
    "@reference"?: string;
}

export interface PositionLocation extends Location {
    "@class": "apex.jorje.data.PositionLocation";
    "@id"?: string;
    "@reference"?: string;
}
// ...
export interface RuntimeException extends Exception {
    "@class": "java.lang.RuntimeException" | "apex.jorje.services.exception.InternalException" | "apex.jorje.services.exception.ParseException";
}
// ...
export interface Stmnt {
    "@class": "apex.jorje.data.ast.Stmnt$BlockStmnt" | "apex.jorje.data.ast.Stmnt$BreakStmnt" | "apex.jorje.data.ast.Stmnt$ContinueStmnt" | "apex.jorje.data.ast.Stmnt$DmlDeleteStmnt" | "apex.jorje.data.ast.Stmnt$DmlInsertStmnt" | "apex.jorje.data.ast.Stmnt$DmlMergeStmnt" | "apex.jorje.data.ast.Stmnt$DmlUndeleteStmnt" | "apex.jorje.data.ast.Stmnt$DmlUpdateStmnt" | "apex.jorje.data.ast.Stmnt$DmlUpsertStmnt" | "apex.jorje.data.ast.Stmnt$DoLoop" | "apex.jorje.data.ast.Stmnt$ExpressionStmnt" | "apex.jorje.data.ast.Stmnt$ForLoop" | "apex.jorje.data.ast.Stmnt$IfElseBlock" | "apex.jorje.data.ast.Stmnt$ReturnStmnt" | "apex.jorje.data.ast.Stmnt$RunAsBlock" | "apex.jorje.data.ast.Stmnt$SwitchStmnt" | "apex.jorje.data.ast.Stmnt$ThrowStmnt" | "apex.jorje.data.ast.Stmnt$TryCatchFinallyBlock" | "apex.jorje.data.ast.Stmnt$VariableDeclStmnt" | "apex.jorje.data.ast.Stmnt$WhileLoop";
}
// ...
export interface UserError {
    "@class": "apex.jorje.data.errors.UserError$Lexical" | "apex.jorje.data.errors.UserError$Syntax";
}